Ensure Content Block views are always loaded (#1018)
Previous behaviour was to rely on manually importing a block for it to
be included (eg; `unordered-list` imports the `list-item` block). While
this is currently always the case, it doesn't necessarily have to be.

One example might be a 3rd party block which wants to create
heading+subheading pair, and set a dependency on the `heading` block
without ever importing it (instead; referencing it by its type name from
within the 3rd party block).

By surfacing the type on the server / during webpack build we're able to
do a couple more checks such as for duplicates and also replace the
sparse config array with a keyed object.

Finally, we can pull up the default `paragraph` block into the pre-build
step and remove that logic from the client code, making it a bit
cleaner.