Expose result of running relationship operations (create/connect/disconnect) (#1033)
* Expose result of running relationship operations (create/connect/disconnect)

Enables future code to inspect the result of specific operations rather
than relying solely on a flattened array of IDs.

In particular; the Content Editor serializes data from slate.js into a
set of operations which it then needs to inject back into the slate.js
document. It does this by tracking `create`/`connect` operation indicies
but without this change was not able to map those indicies back to the
final ID.

The added complication here was the possibility of `null` results due to
Access Control or Where clauses which filtered out items. Now, those
`null` results are preserved in the operation outputs, and filtered by a
second stage function for when obtaining just the flattened ID array is
the desired goal.

* Add edge case tests for mongo specific relationships

* Add changeset