Add queryLimits to List API with maxResults config
This limits the number of results returned in a (sub)query for the List.

Note that the limit is per subquery, not global.  It's intended that
devs base limits on the "reasonable" number of results for the list
type, rather than try to model the capacity of the server.  Combined
with limits on query nesting (not yet implemented), it should be enough
to stop easy OOM DoS attacks.