Fix error "$in requires an array as a second argument, found: missing" when using mongo (#1844)
Adding a new Relationship field when using the Mongoose adapter will
return the value of that field as `null` for any existing items. All new
items created will return the value of the field as `[]`.

This fix prevents `null` causing errors during an aggregation pipeline
where it's expecting an array to search for values within.

To reproduce:

    yarn create keystone-app

Select "blank" / "mongoose" options.

Setup a single list:

    keystone.createList('Group', {
      fields: {
        name: { type: Text },
      }
    });

Then create a new item:

    mutation newGroup {
      createGroup(data:{ name:"foo" }) { id }
    }

Now add a new list and a Relationship field:

    keystone.createList('Group', {
      fields: {
        name: { type: Text },
        labels: { type: Relationship, many: true, ref: 'Label' },
      }
    });

    keystone.createList('Label', {
      fields: {
        name: { type: Text },
      }
    });

Then attempt to query the `id` saved earlier:

    query getGroups {
      allGroup(where: { labels_some: { name: "foo" } }) {
        id
        labels {
          id
        }
      }
    }

Note the error printed to the console:

    $in requires an array as a second argument, found: missing

There is a variation of this error which also occurs when doing a count,
which this commit also fixes.