Add account config params in destroy method (#3220)
* Add account config params in destroy method

We are not setting global configuration (api_key, api_secret, and
cloud_name) in our cloudinary SDK. We are relying on passing these
mandatory config settings in the options param of the upload API.

But in the case of destroy method, we are omitting them and passing on
default empty options object as an argument. This results in a rejected
promise.

To fix this issue, we are now injecting these values in the provided
options object. User can still override these values if required.

Resolves: #2821

* Update .changeset/cyan-keys-sniff.md

* Update .changeset/cyan-keys-sniff.md

Co-authored-by: Aman Singh <aman.singh@thinkmill.com.au>
Co-authored-by: Mike <mike@madebymike.com.au>